requirements.txt
Pyrogram==1.1.9
requests
tgcrypto
pymongo[srv]
python-dotenv
bs4
natsort
Pillow
Image
